+++ binance/client.py
@@ -119,7 +119,8 @@ class BaseClient(ABC):
         m = hmac.new(self.API_SECRET.encode('utf-8'), query_string.encode('utf-8'), hashlib.sha256)
         return m.hexdigest()
 
-    def _order_params(self, data):
+    @staticmethod
+    def _order_params(data):
         """Convert params to list with signature as last element
 
         :param data:
@@ -151,19 +152,20 @@ class BaseClient(ABC):
         data = kwargs.get('data', None)
         if data and isinstance(data, dict):
             kwargs['data'] = data
-        if signed:
-            # generate signature
-            kwargs['data']['timestamp'] = str(int(time.time() * 1000))
-            kwargs['data']['signature'] = self._generate_signature(kwargs['data'])
 
-        # sort get and post params to match signature order
-        if data:
             # find any requests params passed and apply them
             if 'requests_params' in kwargs['data']:
                 # merge requests params into kwargs
                 kwargs.update(kwargs['data']['requests_params'])
                 del(kwargs['data']['requests_params'])
 
+        if signed:
+            # generate signature
+            kwargs['data']['timestamp'] = int(time.time() * 1000)
+            kwargs['data']['signature'] = self._generate_signature(kwargs['data'])
+
+        # sort get and post params to match signature order
+        if data:
             # sort post params
             kwargs['data'] = self._order_params(kwargs['data'])
 
@@ -199,6 +201,7 @@ class Client(BaseClient):
         response = getattr(self.session, method)(uri, **kwargs)
         return self._handle_response(response)
 
+    @staticmethod
     def _handle_response(self, response):
         """Internal helper for handling API responses from the Binance server.
         Raises the appropriate exceptions when necessary; otherwise, returns the
@@ -616,9 +619,9 @@ class Client(BaseClient):
 
         :param symbol: Symbol string e.g. ETHBTC
         :type symbol: str
-        :param start_str: Start date string in UTC format. The iterator will
+        :param start_str: Start date string in UTC format or timestamp in milliseconds. The iterator will
         return the first trade occurring later than this time.
-        :type start_str: str
+        :type start_str: str|int
         :param last_id: aggregate trade ID of the last known aggregate trade.
         Not a regular trade ID. See https://github.com/binance-exchange/binance-official-api-docs/blob/master/rest-api.md#compressedaggregate-trades-list.
 
@@ -641,11 +644,26 @@ class Client(BaseClient):
                 # The difference between startTime and endTime should be less
                 # or equal than an hour and the result set should contain at
                 # least one trade.
-                start_ts = date_to_milliseconds(start_str)
-                trades = self.get_aggregate_trades(
-                    symbol=symbol,
-                    startTime=start_ts,
-                    endTime=start_ts + (60 * 60 * 1000))
+                if type(start_str) == int:
+                    start_ts = start_str
+                else:
+                    start_ts = date_to_milliseconds(start_str)
+                # If the resulting set is empty (i.e. no trades in that interval)
+                # then we just move forward hour by hour until we find at least one
+                # trade or reach present moment
+                while True:
+                    end_ts = start_ts + (60 * 60 * 1000)
+                    trades = self.get_aggregate_trades(
+                        symbol=symbol,
+                        startTime=start_ts,
+                        endTime=end_ts)
+                    if len(trades) > 0:
+                        break
+                    # If we reach present moment and find no trades then there is
+                    # nothing to iterate, so we're done
+                    if end_ts > int(time.time() * 1000):
+                        return
+                    start_ts = end_ts
             for t in trades:
                 yield t
             last_id = trades[-1][self.AGG_ID]
@@ -729,7 +747,8 @@ class Client(BaseClient):
         )
         return kline[0][0]
 
-    def get_historical_klines(self, symbol, interval, start_str, end_str=None):
+    def get_historical_klines(self, symbol, interval, start_str, end_str=None,
+                              limit=500):
         """Get Historical Klines from Binance
 
         See dateparser docs for valid start and end string formats http://dateparser.readthedocs.io/en/latest/
@@ -740,10 +759,12 @@ class Client(BaseClient):
         :type symbol: str
         :param interval: Binance Kline interval
         :type interval: str
-        :param start_str: Start date string in UTC format
-        :type start_str: str
-        :param end_str: optional - end date string in UTC format (default will fetch everything up to now)
-        :type end_str: str
+        :param start_str: Start date string in UTC format or timestamp in milliseconds
+        :type start_str: str|int
+        :param end_str: optional - end date string in UTC format or timestamp in milliseconds (default will fetch everything up to now)
+        :type end_str: str|int
+        :param limit: Default 500; max 1000.
+        :type limit: int
 
         :return: list of OHLCV values
 
@@ -752,13 +773,16 @@ class Client(BaseClient):
         output_data = []
 
         # setup the max limit
-        limit = 500
+        limit = limit
 
         # convert interval to useful value in seconds
         timeframe = interval_to_milliseconds(interval)
 
         # convert our date strings to milliseconds
-        start_ts = date_to_milliseconds(start_str)
+        if type(start_str) == int:
+            start_ts = start_str
+        else:
+            start_ts = date_to_milliseconds(start_str)
 
         # establish first available start timestamp
         first_valid_ts = self._get_earliest_valid_timestamp(symbol, interval)
@@ -767,7 +791,10 @@ class Client(BaseClient):
         # if an end time was passed convert it
         end_ts = None
         if end_str:
-            end_ts = date_to_milliseconds(end_str)
+            if type(end_str) == int:
+                end_ts = end_str
+            else:
+                end_ts = date_to_milliseconds(end_str)
 
         idx = 0
         while True:
@@ -804,6 +831,85 @@ class Client(BaseClient):
                 time.sleep(1)
 
         return output_data
+
+    def get_historical_klines_generator(self, symbol, interval, start_str, end_str=None):
+        """Get Historical Klines from Binance
+
+        See dateparser docs for valid start and end string formats http://dateparser.readthedocs.io/en/latest/
+
+        If using offset strings for dates add "UTC" to date string e.g. "now UTC", "11 hours ago UTC"
+
+        :param symbol: Name of symbol pair e.g BNBBTC
+        :type symbol: str
+        :param interval: Binance Kline interval
+        :type interval: str
+        :param start_str: Start date string in UTC format or timestamp in milliseconds
+        :type start_str: str|int
+        :param end_str: optional - end date string in UTC format or timestamp in milliseconds (default will fetch everything up to now)
+        :type end_str: str|int
+
+        :return: generator of OHLCV values
+
+        """
+
+        # setup the max limit
+        limit = 500
+
+        # convert interval to useful value in seconds
+        timeframe = interval_to_milliseconds(interval)
+
+        # convert our date strings to milliseconds
+        if type(start_str) == int:
+            start_ts = start_str
+        else:
+            start_ts = date_to_milliseconds(start_str)
+
+        # establish first available start timestamp
+        first_valid_ts = self._get_earliest_valid_timestamp(symbol, interval)
+        start_ts = max(start_ts, first_valid_ts)
+
+        # if an end time was passed convert it
+        end_ts = None
+        if end_str:
+            if type(end_str) == int:
+                end_ts = end_str
+            else:
+                end_ts = date_to_milliseconds(end_str)
+
+        idx = 0
+        while True:
+            # fetch the klines from start_ts up to max 500 entries or the end_ts if set
+            output_data = self.get_klines(
+                symbol=symbol,
+                interval=interval,
+                limit=limit,
+                startTime=start_ts,
+                endTime=end_ts
+            )
+
+            # handle the case where exactly the limit amount of data was returned last loop
+            if not len(output_data):
+                break
+
+            # yield data
+            for o in output_data:
+                yield o
+
+            # set our start timestamp using the last value in the array
+            start_ts = output_data[-1][0]
+
+            idx += 1
+            # check if we received less than the required limit and exit the loop
+            if len(output_data) < limit:
+                # exit the while loop
+                break
+
+            # increment next call by our timeframe
+            start_ts += timeframe
+
+            # sleep after every 3rd call to be kind to the API
+            if idx % 3 == 0:
+                time.sleep(1)
 
     def get_ticker(self, **params):
         """24 hour price change statistics.
@@ -1590,6 +1696,163 @@ class Client(BaseClient):
 
         """
         res = self._request_withdraw_api('get', 'accountStatus.html', True, data=params)
+        if not res['success']:
+            raise BinanceWithdrawException(res['msg'])
+        return res
+
+    def get_dust_log(self, **params):
+        """Get log of small amounts exchanged for BNB.
+
+        https://github.com/binance-exchange/binance-official-api-docs/blob/master/wapi-api.md#dustlog-user_data
+
+        :param recvWindow: the number of milliseconds the request is valid for
+        :type recvWindow: int
+
+        :returns: API response
+
+        .. code-block:: python
+
+            {
+                "success": true,
+                "results": {
+                    "total": 2,   //Total counts of exchange
+                    "rows": [
+                        {
+                            "transfered_total": "0.00132256", # Total transfered BNB amount for this exchange.
+                            "service_charge_total": "0.00002699",   # Total service charge amount for this exchange.
+                            "tran_id": 4359321,
+                            "logs": [           # Details of  this exchange.
+                                {
+                                    "tranId": 4359321,
+                                    "serviceChargeAmount": "0.000009",
+                                    "uid": "10000015",
+                                    "amount": "0.0009",
+                                    "operateTime": "2018-05-03 17:07:04",
+                                    "transferedAmount": "0.000441",
+                                    "fromAsset": "USDT"
+                                },
+                                {
+                                    "tranId": 4359321,
+                                    "serviceChargeAmount": "0.00001799",
+                                    "uid": "10000015",
+                                    "amount": "0.0009",
+                                    "operateTime": "2018-05-03 17:07:04",
+                                    "transferedAmount": "0.00088156",
+                                    "fromAsset": "ETH"
+                                }
+                            ],
+                            "operate_time": "2018-05-03 17:07:04" //The time of this exchange.
+                        },
+                        {
+                            "transfered_total": "0.00058795",
+                            "service_charge_total": "0.000012",
+                            "tran_id": 4357015,
+                            "logs": [       // Details of  this exchange.
+                                {
+                                    "tranId": 4357015,
+                                    "serviceChargeAmount": "0.00001",
+                                    "uid": "10000015",
+                                    "amount": "0.001",
+                                    "operateTime": "2018-05-02 13:52:24",
+                                    "transferedAmount": "0.00049",
+                                    "fromAsset": "USDT"
+                                },
+                                {
+                                    "tranId": 4357015,
+                                    "serviceChargeAmount": "0.000002",
+                                    "uid": "10000015",
+                                    "amount": "0.0001",
+                                    "operateTime": "2018-05-02 13:51:11",
+                                    "transferedAmount": "0.00009795",
+                                    "fromAsset": "ETH"
+                                }
+                            ],
+                            "operate_time": "2018-05-02 13:51:11"
+                        }
+                    ]
+                }
+            }
+
+        :raises: BinanceWithdrawException
+
+        """
+        res = self._request_withdraw_api('get', 'userAssetDribbletLog.html', True, data=params)
+        if not res['success']:
+            raise BinanceWithdrawException(res['msg'])
+        return res
+
+    def get_trade_fee(self, **params):
+        """Get trade fee.
+
+        https://github.com/binance-exchange/binance-official-api-docs/blob/master/wapi-api.md#trade-fee-user_data
+
+        :param symbol: optional
+        :type symbol: str
+        :param recvWindow: the number of milliseconds the request is valid for
+        :type recvWindow: int
+
+        :returns: API response
+
+        .. code-block:: python
+
+            {
+                "tradeFee": [
+                    {
+                        "symbol": "ADABNB",
+                        "maker": 0.9000,
+                        "taker": 1.0000
+                    }, {
+                        "symbol": "BNBBTC",
+                        "maker": 0.3000,
+                        "taker": 0.3000
+                    }
+                ],
+                "success": true
+            }
+
+        :raises: BinanceWithdrawException
+
+        """
+        res = self._request_withdraw_api('get', 'tradeFee.html', True, data=params)
+        if not res['success']:
+            raise BinanceWithdrawException(res['msg'])
+        return res
+
+    def get_asset_details(self, **params):
+        """Fetch details on assets.
+
+        https://github.com/binance-exchange/binance-official-api-docs/blob/master/wapi-api.md#asset-detail-user_data
+
+        :param recvWindow: the number of milliseconds the request is valid for
+        :type recvWindow: int
+
+        :returns: API response
+
+        .. code-block:: python
+
+            {
+                "success": true,
+                "assetDetail": {
+                    "CTR": {
+                        "minWithdrawAmount": "70.00000000", //min withdraw amount
+                        "depositStatus": false,//deposit status
+                        "withdrawFee": 35, // withdraw fee
+                        "withdrawStatus": true, //withdraw status
+                        "depositTip": "Delisted, Deposit Suspended" //reason
+                    },
+                    "SKY": {
+                        "minWithdrawAmount": "0.02000000",
+                        "depositStatus": true,
+                        "withdrawFee": 0.01,
+                        "withdrawStatus": true
+                    }
+                }
+            }
+
+        :raises: BinanceWithdrawException
+
+        """
+        res = self._request_withdraw_api('get', 'assetDetail.html', True, data=params)
         if not res['success']:
             raise BinanceWithdrawException(res['msg'])
         return res